Time: 0 ms (100%), Space: 41.1 MB (43.47%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -1,6 +1,6 @@
 class Solution {
     public int reverse(int x) {
-        int rev = 0;
+        long rev = 0;
 
         while(x != 0){
             int lastDigit = x % 10;
